T0: full per-chunk pipeline (DWT->quant->model-zeros->rANS)
chunk.hpp: encode_chunk/decode_chunk wiring DC-removal, 9/7 DWT, dead-zone
  quant, zero-run/EOB token coding (§4.7 model-the-zeros), rANS + bypass.
  Coefficients visited in subband order so zero runs coalesce within dead
  HF bands. Self-contained Payload with serialize/deserialize.
fix(rans): u64 renorm threshold — freq==PROB_SCALE (single-symbol
  alphabet, e.g. an all-zero chunk's lone zero-run) overflowed u32 x_max
  and spun the renorm loop.
test_chunk on scroll-like synthetic: q4 47.2dB/4.5x .. q32 36.0dB/63.6x,
  all-zero chunk 287B, serialize roundtrip lossless.

+++ include/c4d/rans.hpp
@@ -95,7 +95,9 @@ class Encoder {
 public:
     void put(const FreqTable& t, u32 sym) {
         u32 f = t.freq[sym], c = t.cum[sym];
-        u32 x_max = ((RANS_L >> PROB_BITS) << 16) * f;   // renorm threshold
+        // Renorm threshold; u64 because freq up to PROB_SCALE makes this exceed
+        // 2^32 (a single-symbol alphabet hits exactly that).
+        u64 x_max = (u64((RANS_L >> PROB_BITS) << 16)) * f;
         while (state_ >= x_max) { out_.push_back(u16(state_ & 0xffff)); state_ >>= 16; }
         state_ = ((state_ / f) << PROB_BITS) + (state_ % f) + c;
     }
